fix(salt): check bash version is modern, esp macos (#268)
* fix(salt): check bash version is modern, esp macos

* fix(salt): minor refactor; fix unnecessary clones

* fix(salt): macos fixes

+++ salt/salter.sh
@@ -14,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-# Original work at: https://github.com/saltstack-formulas/salter
+# Original work from: https://github.com/saltstack-formulas/salter
 # MODIFIED WORK SECTION has additional copyright under this "License".
 #--------------------------------------------------------------------------
 #
@@ -23,8 +23,8 @@
 #-----------------------------------------------------------------------
 trap exit SIGINT SIGTERM
 [ `id -u` != 0 ] && echo && echo "Run script with sudo, exiting" && echo && exit 1
-declare -A your solution fork || (echo "bash v4 or later is required" && exit 1)
 
+RC=0
 BASE=/srv
 BASE_ETC=/etc
 STATEDIR=''
@@ -32,11 +32,30 @@ if [ `uname` == 'FreeBSD' ]; then
     BASE=/usr/local/etc
     BASE_ETC=/usr/local/etc
     STATEDIR=states
-elif [ `uname` == 'Darwin' ]; then
-    USER=$( stat -f "%Su" /dev/console )
 fi
 PILLARFS=${BASE:-/srv}/pillar
 SALTFS=${BASE:-/srv}/salt/${STATEDIR}
+SKIP_UNNECESSARY_CLONE=''
+HOMEBREW=/usr/local/bin/brew
+
+# bash version must be modern
+declare -A your solution fork 2>/dev/null || RC=$?
+if (( RC > 0 )) && [ "$( uname )" = "Darwin" ]; then
+    echo "[warning] your bash version is too old ..."
+    # macos needs homebrew (unattended https://github.com/Homebrew/legacy-homebrew/issues/46779#issuecomment-162819088)
+    export PY_VER=3
+    export USER=$( stat -f "%Su" /dev/console )
+    export HOMEBREW=/usr/local/bin/brew
+    ${HOMEBREW} >/dev/null 2>&1
+    (( $? == 127 )) && su - ${USER} -c 'echo | /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"'
+    # macos needs modern bash
+    (( RC > 0 )) && (su - ${USER} -c "${HOMEBREW} install bash" || exit 12) && RC=0
+fi
+if (( RC > 0 )); then
+    # linux needs modern bash
+    echo "[error] your bash version is too old ..."
+    exit ${RC}
+fi
 
 #-----------------------------------------
 #   Adaption layer for OS package handling
@@ -58,16 +77,17 @@ pkg-install() {
     PACKAGES=${@}
     case ${OSTYPE} in
     darwin*) for p in ${PACKAGES}; do
-                 su ${USER} -c "brew install ${p}"
-                 su ${USER} -c "brew unlink ${p} 2>/dev/null && brew link ${p} 2>/dev/null"
+                 su - ${USER} -c "${HOMEBREW} install ${p}"
+                 su - ${USER} -c "${HOMEBREW} unlink ${p} 2>/dev/null"
+                 su - ${USER} -c "${HOMEBREW} link ${p} 2>/dev/null"
              done
              awk >/dev/null 2>&1
              if (( $? == 134 )); then
                  ## https://github.com/atomantic/dotfiles/issues/23#issuecomment-298784915 ###
-                 brew uninstall gawk
-                 brew uninstall readline
-                 brew install readline
-                 brew install gawk
+                 su - ${USER} -c "${HOMEBREW} uninstall gawk"
+                 su - ${USER} -c "${HOMEBREW} uninstall readline"
+                 su - ${USER} -c "${HOMEBREW} install readline"
+                 su - ${USER} -c "${HOMEBREW} install gawk"
              fi
              ;;
 
@@ -99,7 +119,7 @@ pkg-update() {
     PACKAGES=${@}
     case ${OSTYPE} in
     darwin*) for p in ${PACKAGES}; do
-                 su ${USER} -c "brew upgrade ${p}"
+                 su - ${USER} -c "${HOMEBREW} upgrade ${p}"
              done
              ;;
     linux*)  if [ -f "/usr/bin/zypper" ]; then
@@ -122,7 +142,7 @@ pkg-remove() {
     PACKAGES=${@}
     case ${OSTYPE} in
     darwin*) for p in ${PACKAGES}; do
-                 su ${USER} -c "brew uninstall ${p} --force"
+                 su - ${USER} -c "${HOMEBREW} uninstall ${p} --force"
              done
              ;;
     linux*)  if [ -f "/usr/bin/zypper" ]; then
@@ -186,22 +206,19 @@ salt-bootstrap() {
              ### https://stackoverflow.com/questions/34386527/symbol-not-found-pycodecinfo-getincrementaldecoder
              su - ${USER} -c 'hash -r python'
 
-             ### Secure install pip https://pip.pypa.io/en/stable/installing/
+             ### Secure install https://pip.pypa.io/en/stable/installing/
              su - ${USER} -c 'curl https://bootstrap.pypa.io/get-pip.py -o ${PWD}/get-pip.py'
              sudo python ${PWD}/get-pip.py 2>/dev/null
 
-             [ -x /usr/local/bin/brew ] | su - ${USER} -c '/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"'
-             su - ${USER} -c '/usr/local/bin/pip install --upgrade wrapper barcodenumber npyscreen'
-             [[ ! -x /usr/local/bin/brew ]] && echo "Install homebrew (https://docs.brew.sh/Installation.html)" && exit 1
-
              /usr/local/bin/salt --version >/dev/null 2>&1
              if (( $? > 0 )); then
-                 su ${USER} -c 'brew install saltstack'
+                 su - ${USER} -c "${HOMEBREW} install saltstack"
              else
-                 su ${USER} -c 'brew upgrade saltstack'
+                 su - ${USER} -c "${HOMEBREW} upgrade saltstack"
              fi
-             su ${USER} -c 'brew unlink saltstack && brew link saltstack'
-             su ${USER} -c 'brew tap homebrew/services'
+             su - ${USER} -c "${HOMEBREW} unlink saltstack"
+             su - ${USER} -c "${HOMEBREW} link saltstack"
+             su - ${USER} -c "${HOMEBREW} tap homebrew/services"
              echo $( hostname ) >/etc/salt/minion_id
              cp /usr/local/etc/saltstack/minion /etc/salt/minion 2>/dev/null
              sed -i"bak" "s/#file_client: remote/file_client: local/" /etc/salt/minion 2>/dev/null
@@ -243,8 +260,8 @@ EOF
         ### Enforce python3
         rm /usr/bin/python 2>/dev/null; ln -s /usr/bin/python3 /usr/bin/python
     fi
-    ### install salt-api (except arch)
-    [ -f "/etc/arch-release" ] || pkg-install salt-api
+    ### install salt-api (except arch/macos)
+    [ ! -f "/etc/arch-release" ] && [ "$(uname)" != "Darwin" ] && pkg-install salt-api
 
     ### salt services
     if [[ "`uname`" == "FreeBSD" ]] || [[ "`uname`" == "Darwin" ]]; then
@@ -265,7 +282,7 @@ EOF
 
 setup-log() {
     LOG=${1}
-    mkdir -p ${solution['logdir']} 2>/dev/null
+    mkdir -p ${solution[logdir]} 2>/dev/null
     salt-call --versions >>${LOG} 2>&1
     [ -f "${PILLARFS}/site.j2" ] && cat ${PILLARFS}/site.j2 >>${LOG} 2>&1
     [ -n "${DEBUGG_ON}" ] && salt-call pillar.items --local >> ${LOG} 2>&1 && echo >>${LOG} 2>&1
@@ -275,6 +292,8 @@ setup-log() {
 }
 
 gitclone() {
+    [ -n "${SKIP_UNNECESSARY_CLONE}" ] && return 0
+
     URI=${1} && ENTITY=${2} && REPO=${3} && ALIAS=${4} && SUBDIR=${5}
     echo "cloning ${REPO} from ${ENTITY} ..."
     rm -fr ${SALTFS}/namespaces/${ENTITY}/${REPO} 2>/dev/null
@@ -295,13 +314,11 @@ highstate() {
 
     ## prepare states
     ACTION=${1} && STATEDIR=${2} && TARGET=${3}
-    for PROFILE in ${solution[states]}/${ACTION}/${TARGET} ${your[states]}/${ACTION}/${TARGET}
+    for PROFILE in ${solution[saltdir]}/${ACTION}/${TARGET} ${your[saltdir]}/${ACTION}/${TARGET}
     do  
-        set -xv
         [ -f ${PROFILE}.sls ] && cp ${PROFILE}.sls ${SALTFS}/top.sls && break
         [ -f ${PROFILE}/init.sls ] && cp ${PROFILE}/init.sls ${SALTFS}/top.sls && break
     done
-    [ -z "${DEBUGG_ON}" ] && set +xv
     [ ! -f ${SALTFS}/top.sls ] && echo "Failed to find ${TARGET}.sls or ${TARGET}/init.sls" && usage
 
     ## prepare pillars
@@ -327,7 +344,7 @@ highstate() {
          salt)       continue;;                    ##already cloned?
          *)          source=${formula} ;;
          esac
-         gitclone 'https://github.com' saltstack-formulas ${source}-formula ${formula} ${formula}
+         gitclone 'https://github.com' "${solution[provider]}" ${source}-formula ${formula} ${formula}
     done
 
     ## run states
@@ -393,12 +410,12 @@ while getopts ":i:l:r:u:" option; do
 done
 shift $((OPTIND-1))
 
-business-logic() {
+salter-engine() {
     ## remove option
     if [ "${ACTION}" == 'remove' ] && [ -n "${TARGET}" ]; then
         echo "${solution[targets]}" | grep "${TARGET}" >/dev/null 2>&1
-        if (( $? == 0 )) || [ -f ${solution[states]}/${ACTION}/${TARGET}.sls ]; then
-           highstate remove ${solution[states]} ${TARGET}
+        if (( $? == 0 )) || [ -f ${solution[saltdir]}/${ACTION}/${TARGET}.sls ]; then
+           highstate remove ${solution[saltdir]} ${TARGET}
            return 0
         fi
     fi
@@ -407,95 +424,100 @@ business-logic() {
     case "${TARGET}" in
     bootstrap)  salt-bootstrap ;;
 
-    menu)       pip install --pre wrapper barcodenumber npyscreen || exit 1
-                ([ -x ${SALTFS}/contrib/menu.py ] && ${SALTFS}/contrib/menu.py ${solution[states]}/install) || exit 2
-                highstate install ${solution[states]} ${TARGET} ;;
+    menu)       pip${PY_VER} install --pre wrapper barcodenumber npyscreen || exit 1
+                ([ -x ${SALTFS}/contrib/menu.py ] && ${SALTFS}/contrib/menu.py ${solution[saltdir]}/install) || exit 2
+                highstate install "${solution[saltdir]}" ${TARGET} ;;
 
-    salter)     gitclone 'https://github.com' saltstack-formulas salt-formula salt salt
+    salter)     gitclone 'https://github.com' "${solution[provider]}" salt-formula salt salt
                 gitclone ${solution[uri]} ${solution[entity]} ${solution[repo]} ${solution[alias]} ${solution[subdir]}
-                highstate install ${solution[states]} salt
+                highstate install "${solution[saltdir]}" salt
                 rm /usr/local/bin/salter.sh 2>/dev/null
                 ln -s ${solution[homedir]}/salter.sh /usr/local/bin/salter.sh ;;
 
     ${solution[alias]})
-                solution-tasks ${solution[alias]} ;;
+                custom-install ${solution[alias]} ;;
 
-    *)          ## profiles (STATES/FORMULAS)
+    *)          ## PROFILES (STATES/FORMULAS)
                 echo "${solution[targets]}" | grep "${TARGET}" >/dev/null 2>&1
-                if (( $? == 0 )) || [ -f ${solution[states]}/install/${TARGET}.sls ]; then
-                    highstate install ${solution[states]} ${TARGET}
-                    optional-post-install-work ${TARGET}
+                if (( $? == 0 )) || [ -f ${solution[saltdir]}/install/${TARGET}.sls ]; then
+                    highstate install ${solution[saltdir]} ${TARGET}
+                    custom-postinstall ${TARGET}
                 fi
     esac
 }
 
 #########################################################################
-#
-# MODIFIED WORK SECTION
-# Copyright 2019 The OpenSDS Authors 
-#
+# SOLUTION: Copyright 2019 The OpenSDS Authors
 #########################################################################
 
-mandatory-solution-repo-description() {
-    ### repo details ###
+developer-definitions() {
+    fork['uri']="https://github.com"
+    fork['entity']="noelmcloughlin"
+    fork['branch']="fixes"
+    fork['solutions']="salter packages-formula golang-formula postgres-formula"
+}
+
+solution-definitions() {
+    ### solution polyrepo or monorepo ###
     solution['saltmaster']=""
     solution['uri']="https://github.com"
     solution['entity']="opensds"
     solution['repo']="opensds-installer"
     solution['alias']="opensds"
-    solution['targets']="opensds|gelato|auth|hotpot|backend|dashboard|database|dock|keystone|config|infra|sushi|freespace|telemetry|deepsea"
     solution['subdir']="salt"
+    solution['provider']="saltstack-formulas"
+    solution['profiles']="opensds|gelato|auth|hotpot|backend|dashboard|database|dock|keystone|config|infra|sushi|freespace|telemetry|deepsea"
 
-    ### giving these values ###
-    solution['homedir']="${SALTFS}/namespaces/${solution['entity']}/${solution[repo]}/${solution[subdir]}"
-    solution['states']="${solution[homedir]}/file_roots"
+    ### derivatives
+    solution['homedir']="${SALTFS}/namespaces/${solution[entity]}/${solution[repo]}/${solution[subdir]}"
+    solution['saltdir']="${solution[homedir]}/file_roots"
     solution['pillars']="${solution[homedir]}/pillar_roots"
     solution['logdir']="/tmp/${solution[entity]}-${solution[repo]}"
 
-    ### YOUR STUFF HERE ###
-    your['states']="${SALTFS}/namespaces/your/file_roots"
+    your['saltdir']="${SALTFS}/namespaces/your/file_roots"
     your['pillars']="${SALTFS}/namespaces/your/pillar_roots"
-
-    mkdir -p ${solution[states]} ${solution[pillars]} ${your[states]} ${your[pillars]} ${solution[logdir]} ${PILLARFS} ${BASE_ETC}/salt 2>/dev/null
+    mkdir -p ${solution[saltdir]} ${solution[pillars]} ${your[saltdir]} ${your[pillars]} ${solution[logdir]} ${PILLARFS} ${BASE_ETC}/salt 2>/dev/null
 }
 
-optional-developer-settings() {
-    fork['uri']="https://github.com"
-    fork['entity']="noelmcloughlin"
-    fork['branch']="fixes"
-    fork['solutions']="opensds-installer salter packages-formula golang-formula"
-}
-
-solution-tasks() {
+custom-install() {
+    ### required - salter-engine is insufficient ###
     gitclone ${solution[uri]} ${solution[entity]} ${solution[repo]} ${solution[alias]} ${solution[subdir]}
+    SKIP_UNNECESSARY_CLONE='true'
     losetup -D 2>/dev/null
-    highstate install ${solution[states]} infra
-    highstate install ${solution[states]} telemetry
-    highstate install ${solution[states]} keystone
+    highstate install ${solution[saltdir]} infra
+    highstate install ${solution[saltdir]} telemetry
+    highstate install ${solution[saltdir]} keystone
     #show-logger /tmp/devstack/stack.sh.log
-    highstate install ${solution[states]} config
-    highstate install ${solution[states]} database
-    highstate install ${solution[states]} auth
-    highstate install ${solution[states]} hotpot
-    highstate install ${solution[states]} sushi
-    highstate install ${solution[states]} backend
-    highstate install ${solution[states]} dock
-    highstate install ${solution[states]} dashboard
-    highstate install ${solution[states]} gelato
-    highstate install ${solution[states]} freespace
+    highstate install ${solution[saltdir]} config
+    highstate install ${solution[saltdir]} database
+    highstate install ${solution[saltdir]} auth
+    highstate install ${solution[saltdir]} hotpot
+    highstate install ${solution[saltdir]} sushi
+    highstate install ${solution[saltdir]} backend
+    highstate install ${solution[saltdir]} dock
+    highstate install ${solution[saltdir]} dashboard
+    highstate install ${solution[saltdir]} gelato
+    highstate install ${solution[saltdir]} freespace
     cp ${SALTFS}/namespaces/${solution['entity']}/${solution[repo]}/conf/*.json /etc/opensds/ 2>/dev/null
 }
 
-optional-post-install-work(){
-    ## SUSE/Deepsea
+custom-postinstall() {
+    LXD=${SALTFS}/namespaces/saltstack-formulas/lxd-formula
+    # see https://github.com/saltstack-formulas/lxd-formula#clone-and-symlink
+    [ -d "${LXD}/_modules" ] && ln -s ${LXD}/_modules ${SALTFS}/_modules 2>/dev/null
+    [ -d "${LXD}/_states" ] && ln -s ${LXD}/_states ${SALTFS}/_states 2>/dev/null
+
+    # SUSE/Deepsea/Ceph
     if (( $? == 0 )) && [[ "${1}" == "deepsea" ]]; then
        salt-call --local grains.append deepsea default ${solution['saltmaster']}
        cp ${solution['homedir']}/file_roots/install/deepsea_post.sls ${SALTFS}/${STATES_DIR}/top.sls
     fi
 }
 
-## MAIN ##
+### MAIN
 
-optional-developer-settings
-mandatory-solution-repo-description
-business-logic
+developer-definitions
+solution-definitions
+salter-engine
+exit $?
+
